Complete the windows/filesystem rename so main builds
The windows/ms_fscc -> windows/filesystem rename merged in #535 was only
partial: the directory and doc.go/infoclass were renamed, but the struct and
test files still declared package ms_fscc and the SMB2 client still imported
windows/ms_fscc. The result did not compile (found packages filesystem and
ms_fscc in the same directory; no package windows/ms_fscc).

Finish the rename: declare package filesystem in file_query_information.go,
file_set_information.go, fs_information.go, and filesystem_test.go; point the
smb_v20 client's info.go and filemgmt.go imports and references at
windows/filesystem; and correct the stale package name in the infoclass doc
comment.

go build ./... and the windows/filesystem and smb_v20 client tests pass.
