+++ windows/filesystem/file_set_information.go
@@ -1,4 +1,4 @@
-package ms_fscc
+package filesystem
 
 import (
 	"encoding/binary"
